Fixed up the ef tool to be a bit more organized.
Adding C++ classes for effort datasets -- todo merge these with the various python classes libra uses (i.e. wrap things and use wrapped version in python)
Adding clustering tools for Chapter 6 of dissertation.

+++ ClusterDataSet.C
@@ -71,7 +71,7 @@ ClusterDataSet* ClusterDataSet::randomSubset(unsigned size) const {
     unsigned m = size;
     
     while (m > 0) {
-        if (rnGenerator(remaining) < m) {
+      if (rnGenerator(remaining) < (int)m) {
             *out = first;
             ++out;
             --m;
